refactor(data/product): mockserver를 구현함에 따라서 sqlite를 사용하지 않기 때문에 생성인자에 context 삭제하기
