+++ app/src/main/java/woowacourse/shopping/data/product/datasource/ProductDataSourceImpl.kt
@@ -1,12 +1,10 @@
 package woowacourse.shopping.data.product.datasource
 
-import android.content.Context
 import woowacourse.shopping.data.product.ProductEntity
 import woowacourse.shopping.data.product.remote.ProductRemote
 import woowacourse.shopping.data.product.remote.ProductRemoteImpl
 
 class ProductDataSourceImpl(
-    context: Context,
     private val productRemote: ProductRemote = ProductRemoteImpl(),
 ) : ProductDataSource {
