+++ app/src/main/java/woowacourse/shopping/shopping/ShoppingActivity.kt
@@ -14,10 +14,10 @@ import woowacourse.shopping.data.recentviewed.repository.RecentViewedProductRepo
 import woowacourse.shopping.databinding.ActivityShoppingBinding
 import woowacourse.shopping.model.ProductUiModel
 import woowacourse.shopping.model.RecentViewedProductUiModel
+import woowacourse.shopping.shopping.adapter.ProductItemDecoration
 import woowacourse.shopping.shopping.adapter.ShoppingProductCountPicker
 import woowacourse.shopping.shopping.adapter.ShoppingRecyclerAdapter
 import woowacourse.shopping.shopping.adapter.ShoppingRecyclerSpanSizeManager
-import woowacourse.shopping.shopping.adapter.viewholder.ProductItemDecoration
 import woowacourse.shopping.shopping.navigator.ShoppingNavigator
 import woowacourse.shopping.shopping.navigator.ShoppingNavigatorImpl
 
@@ -28,7 +28,7 @@ class ShoppingActivity : AppCompatActivity(), ShoppingContract.View {
     private val presenter: ShoppingContract.Presenter by lazy {
         ShoppingPresenter(
             view = this,
-            productRepository = ProductRepositoryImpl(this),
+            productRepository = ProductRepositoryImpl(),
             cartRepository = CartRepositoryImpl(this),
             recentViewedProductRepository = RecentViewedProductRepositoryImpl(this)
         )
